// src/types/enums.ts
export const enum UIAppType {
  webBased = 'webBased',
  extension = 'extension',
  outlook = 'outlook',
  expanded = 'expanded',
}

export const enum UISize {
  sm = 'sm',
  md = 'md',
  lg = 'lg',
  xl = 'xl',
}

export const enum UIColor {
  primary = 'primary',
  secondary = 'secondary',
  special = 'special',
}

export const enum UIPlacement {
  center = 'center',
  top = 'top',
  topEnd = 'topEnd',
  topStart = 'topStart',
  left = 'left',
  leftEnd = 'leftEnd',
  leftStart = 'leftStart',
  right = 'right',
  rightEnd = 'rightEnd',
  rightStart = 'rightStart',
  bottom = 'bottom',
  bottomEnd = 'bottomEnd',
  bottomStart = 'bottomStart',
}

export const enum UIModalButton {
  confirm = 'confirm',
  delete = 'delete',
  buyNow = 'buyNow',
  cancel = 'cancel',
}

export const enum UILibraryView {
  list = 'list',
  grid = 'grid',
}